refactor(launcher): extract pre-Compose pipeline into LauncherBootstrap
Main.kt was carrying ~100 lines of JVM-process setup before the
Compose application{} call: logs-dir property, DataDirMover,
sessionId, ActionRing startup line, NetworkState init, X11 hack,
display diag, Files.createDirectories, crash-reporter wiring,
single-instance lock, migration detection, startKoin. None of
that is UI-shaped; consolidating it into LauncherBootstrap.preBoot()
gives a single entry that returns the few values (paths,
pendingMigration, crashReporter) the Compose layer actually
needs.

preBoot() accepts an extraModules: List<Module> param so the ui
module can register its Koin singletons (SkinManager, GameConsole)
into the same context without the launcher needing to depend on
ui types. PuppetServerLoader.startIfRequested() stays in Main --
it's client-ui-housed and must run after Koin + before
application{}.

Captured boot.pendingMigration into a local val at the MigrationScreen
call site -- a public property in a different module doesn't carry
the if-not-null smart cast across the nested call boundary.

Main.kt 770 -> 651 LOC. Unused imports dropped (CrashReporter,
DataDirMigration, DataDirMover, PlatformPaths, SingleInstance,
appModule, networkModule, Files, UUID, startKoin, exitProcess,
XToolkitOverride+DisplayDiagnostics call-site only).
